requirements.txt
dynaconf==3.2.1
pydantic==1.8.1
uvicorn==0.23.2
fastapi==0.63.0
starlette==0.13.6
greenlet==2.0.2
gevent==23.7.0
dash>=2.5.1
numpy==1.22.4
Pillow==10.0.1
numpy==1.22.4
dash-bootstrap-components>=1.1.0
dash-mantine-components>=0.10.1
dash-iconify>=0.1.2
dash-extensions>=0.1.6
kafka-python==2.0.2
opentelemetry-sdk==1.19.0
opentelemetry-distro==0.40b0
opentelemetry-exporter-otlp-proto-grpc==1.19.0
opentelemetry-instrumentation-fastapi==0.40b0
opentelemetry-instrumentation-kafka-python==0.40b0
opentelemetry-exporter-jaeger==1.19.0